deallocate resource at instance.remove
